Generalize `sub` into `higher_ranked`
