+++ src/librustc/middle/traits/select.rs
@@ -24,7 +24,6 @@ use super::{util};
 use middle::mem_categorization::Typer;
 use middle::subst::{Subst, Substs, VecPerParamSpace};
 use middle::ty;
-use middle::typeck::check::regionmanip;
 use middle::typeck::infer;
 use middle::typeck::infer::{InferCtxt, TypeSkolemizer};
 use middle::ty_fold::TypeFoldable;
@@ -1703,8 +1702,8 @@ impl<'cx, 'tcx> SelectionContext<'cx, 'tcx> {
         // FIXME(pcwalton): This is a bogus thing to do, but
         // it'll do for now until we get the new trait-bound
         // region skolemization working.
-        let (_, new_signature) =
-            regionmanip::replace_late_bound_regions(
+        let (new_signature, _) =
+            ty::replace_late_bound_regions(
                 self.tcx(),
                 closure_type.sig.binder_id,
                 &closure_type.sig,
@@ -1712,14 +1711,16 @@ impl<'cx, 'tcx> SelectionContext<'cx, 'tcx> {
                          infer::LateBoundRegion(obligation.cause.span, br)));
 
         let arguments_tuple = new_signature.inputs[0];
-        let trait_ref = Rc::new(ty::TraitRef {
-            def_id: obligation.trait_ref.def_id,
-            substs: Substs::new_trait(
+        let substs =
+            Substs::new_trait(
                 vec![arguments_tuple.subst(self.tcx(), substs),
                      new_signature.output.unwrap().subst(self.tcx(), substs)],
                 vec![],
                 vec![],
-                obligation.self_ty())
+                obligation.self_ty());
+        let trait_ref = Rc::new(ty::TraitRef {
+            def_id: obligation.trait_ref.def_id,
+            substs: substs,
         });
 
         self.confirm(obligation.cause,
