+++ src/librustc/middle/ty.rs
@@ -5598,3 +5598,33 @@ impl AutoDerefRef {
         self.autoderefs == 0 && self.autoref.is_none()
     }
 }
+
+pub fn replace_late_bound_regions<T>(
+    tcx: &ty::ctxt,
+    binder_id: ast::NodeId,
+    value: &T,
+    mapf: |ty::BoundRegion| -> ty::Region)
+    -> (T, HashMap<ty::BoundRegion,ty::Region>)
+    where T : TypeFoldable + Repr
+{
+    debug!("replace_late_bound_regions({})", value.repr(tcx));
+
+    let mut map = HashMap::new();
+    let value = {
+        let mut f = ty_fold::RegionFolder::regions(tcx, |r| {
+            debug!("region r={}", r.to_string());
+            match r {
+                ty::ReLateBound(s, br) if s == binder_id => {
+                    * match map.entry(br) {
+                        Vacant(entry) => entry.set(mapf(br)),
+                        Occupied(entry) => entry.into_mut(),
+                    }
+                }
+                _ => r
+            }
+        });
+        value.fold_with(&mut f)
+    };
+    debug!("resulting map: {}", map);
+    (value, map)
+}
